Support transparency map from G3DJ spec, refactored elements to rely less on ScreenManager's singleton

+++ Framework/Screens/InputElements/ListSelector.cs
@@ -47,7 +47,7 @@ namespace Spectrum.Framework.Screens.InputElements
         public override void PositionUpdate()
         {
             base.PositionUpdate();
-            int distFromBottom = Manager.Root.TotalHeight - (Y.Flat + TotalHeight);
+            int distFromBottom = Manager.Root.TotalHeight - ((int)Y.Flat + TotalHeight);
             if (distFromBottom < 0)
                 Y.Flat += distFromBottom;
         }
